requirements.txt
# Core PDF and text extraction
pdfplumber==0.11.4

# Embedding model (downloads all-MiniLM-L6-v2 automatically on first run)
sentence-transformers==3.3.1

# Vector similarity search
faiss-cpu==1.9.0

# Google Gemini API client
google-generativeai==0.8.4

# Numerical computing (required by FAISS and sentence-transformers)
numpy==1.26.4

# Load environment variables from .env file
python-dotenv==1.0.1
